Changed the type of the "albedo" render target from `uint8` to `float`. Because render targets with different bitplanes are not allowed in WebGL.

+++ example/deferred_shading.js
@@ -17,7 +17,7 @@ camera.zoom(500.0) // 10.0
 
 const fbo = regl.framebuffer({
   color: [
-    regl.texture({type: 'uint8'}), // albedo
+    regl.texture({type: 'float'}), // albedo
     regl.texture({type: 'float'}), // normal
     regl.texture({type: 'float'}) // position
   ],
